server/gemini/requirements.txt
# Run install.sh to install (uses dlib-bin to avoid dlib build on macOS)
fastapi==0.111.0
uvicorn[standard]==0.29.0
google-generativeai==0.7.2
face-recognition-models>=0.3.0
# face-recognition==1.3.0  # installed by install.sh with --no-deps
numpy==1.26.4
Pillow==10.3.0
python-dotenv==1.0.1
pydantic==2.7.1
